Purge epee/local_ip.h
