+++ contrib/epee/src/net_utils_base.cpp
@@ -4,7 +4,55 @@
 #include <boost/uuid/uuid_io.hpp>
 
 #include "epee/string_tools.h"
-#include "epee/net/local_ip.h"
+#include "epee/int-util.h"
+
+namespace {
+
+    bool is_ip_loopback(uint32_t ip)
+    {
+      ip = SWAP32LE(ip);
+      if ((ip & 0xff) == 0x7f) // 127.0.0.0/8
+        return true;
+
+      return false;
+    }
+
+    bool is_ip_local(uint32_t ip)
+    {
+      ip = SWAP32LE(ip);
+      /*
+      private network ranges:
+      10.0.0.0/8
+      172.16.0.0/12
+      192.168.0.0/16
+
+      carrier-grade NAT network range:
+      100.64.0.0/10
+
+      link-local addresses:
+      169.254.0.0/16
+      */
+
+      // Extremely bizarrely, IPs are stored in little-endian order in epee, which is just plain
+      // wrong, but we have to deal with.  (Jason)
+      if ((ip & 0xff) == 0x0a) // 10.0.0.0/8
+        return true;
+
+      if ((ip & 0xf0ff) == 0x10ac) // 172.16.0.0/12 (0xf0ff looks strange because of the little endian nonsense)
+        return true;
+
+      if ((ip & 0xffff) == 0xa8c0) // 192.168.0.0/16
+        return true;
+
+      if ((ip & 0xc0ff) == 0x4064) // 100.64.0.0/10
+        return true;
+
+      if ((ip & 0xffff) == 0xfea9) // 169.254.0.0/16
+        return true;
+
+      return false;
+    }
+}
 
 namespace epee { namespace net_utils
 {
@@ -18,8 +66,8 @@ namespace epee { namespace net_utils
 	{ return string_tools::get_ip_string_from_int32(ip()) + ":" + std::to_string(port()); }
 
 	std::string ipv4_network_address::host_str() const { return string_tools::get_ip_string_from_int32(ip()); }
-	bool ipv4_network_address::is_loopback() const { return net_utils::is_ip_loopback(ip()); }
-	bool ipv4_network_address::is_local() const { return net_utils::is_ip_local(ip()); }
+	bool ipv4_network_address::is_loopback() const { return is_ip_loopback(ip()); }
+	bool ipv4_network_address::is_local() const { return is_ip_local(ip()); }
 
 	bool ipv6_network_address::equal(const ipv6_network_address& other) const noexcept
 	{ return is_same_host(other) && port() == other.port(); }
@@ -45,8 +93,8 @@ namespace epee { namespace net_utils
 	{ return string_tools::get_ip_string_from_int32(subnet()) + "/" + std::to_string(m_mask); }
 
 	std::string ipv4_network_subnet::host_str() const { return string_tools::get_ip_string_from_int32(subnet()) + "/" + std::to_string(m_mask); }
-	bool ipv4_network_subnet::is_loopback() const { return net_utils::is_ip_loopback(subnet()); }
-	bool ipv4_network_subnet::is_local() const { return net_utils::is_ip_local(subnet()); }
+	bool ipv4_network_subnet::is_loopback() const { return is_ip_loopback(subnet()); }
+	bool ipv4_network_subnet::is_local() const { return is_ip_local(subnet()); }
 	bool ipv4_network_subnet::matches(const ipv4_network_address &address) const
 	{
 		return (address.ip() & ~(0xffffffffull << m_mask)) == subnet();
