+++ src/cryptonote_core/service_node_list.cpp
@@ -39,7 +39,6 @@ extern "C" {
 }
 
 #include "ringct/rctSigs.h"
-#include "epee/net/local_ip.h"
 #include "cryptonote_tx_utils.h"
 #include "cryptonote_basic/tx_extra.h"
 #include "cryptonote_basic/hardfork.h"
@@ -3104,7 +3103,7 @@ namespace service_nodes
       }
     }
 
-    if (!debug_allow_local_ips && !epee::net_utils::is_ip_public(proof->public_ip))
+    if (!debug_allow_local_ips && !tools::net_utils::is_ip_public(tools::net_utils::from_big_endian(proof->public_ip)))
       REJECT_PROOF("public_ip is not actually public");
 
     //
