+++ src/p2p/net_peerlist.h
@@ -46,7 +46,6 @@
 #include "crypto/crypto.h"
 #include "cryptonote_config.h"
 #include "epee/net/enums.h"
-#include "epee/net/local_ip.h"
 #include "p2p_protocol_defs.h"
 #include "common/random.h"
 #include "common/fs.h"
